Replace MIO0 by SMSR00 compression
Encoder by @klorfmorf

Two files left in the partial source from bbgames reveal that SM64 used at one point "SMSR00" compression:
- data/tanidata.szp
- map/stage99/Gstage1.szp

slidec.s in the MK64 source has 2 versions of the function, the latter for decoding MIO0, and the first for SMSR00. Only the latter is used.

In SM64, slidec.s is identical to the second function in MK64's slidec.s, except the name of the function was changed, but NOTHING else. The gotos still have 2 at the end of them.

This was likely used at one point in SM64's development, but when was it switched to MIO0? The only way to find out would be to look at the display lists in Gstage1.szp.